Migrated submission modules to SubmissionActivityModule, and included in MainActivityModule
